+++ packages/lib/src/globalConfig/globalConfig.ts
@@ -25,12 +25,12 @@ export enum ModelAutoTypeCheckingMode {
  */
 export interface GlobalConfig {
   /**
-   * Model auto type-checkig mode.
+   * Model auto type-checking mode.
    */
   modelAutoTypeChecking: ModelAutoTypeCheckingMode
 
   /**
-   * ID generator function for $modelId.
+   * ID generator function for model ids.
    */
   modelIdGenerator(): string