+++ packages/lib/src/model/Model.ts
@@ -1,8 +1,10 @@
-import { AnyModelProp } from "model"
 import { ActionContextActionType } from "../action/context"
 import { wrapInAction } from "../action/wrapInAction"
+import { getGlobalConfig } from "../globalConfig"
 import { memoTransformCache } from "../propTransform/propTransform"
 import { typesObject } from "../typeChecking/object"
+import { typesString } from "../typeChecking/primitives"
+import { tProp } from "../typeChecking/tProp"
 import { LateTypeChecker } from "../typeChecking/TypeChecker"
 import { typesUnchecked } from "../typeChecking/unchecked"
 import { assertIsObject, failure, propNameToSetterActionName } from "../utils"
@@ -12,20 +14,24 @@ import {
   baseModelPropNames,
   ModelClass,
   modelInitializedSymbol,
-  ModelInstanceData,
+  ModelInstanceData
 } from "./BaseModel"
+import { ModelMetadata } from "./getModelMetadata"
 import { modelIdKey, modelTypeKey } from "./metadata"
 import { modelInitializersSymbol } from "./modelClassInitializer"
 import { ModelConstructorOptions } from "./ModelConstructorOptions"
 import { getInternalModelClassPropsInfo, setInternalModelClassPropsInfo } from "./modelPropsInfo"
-import { modelDataTypeCheckerSymbol, modelUnwrappedClassSymbol } from "./modelSymbols"
+import { modelMetadataSymbol, modelUnwrappedClassSymbol } from "./modelSymbols"
 import {
+  AnyModelProp,
+  idProp,
   ModelProps,
   ModelPropsToInstanceCreationData,
   ModelPropsToInstanceData,
   ModelPropsToPropsCreationData,
   ModelPropsToPropsData,
   ModelPropsToSetterActions,
+  prop
 } from "./prop"
 import { assertIsModelClass } from "./utils"
 
@@ -49,22 +55,38 @@ export interface _Model<SuperModel, TProps extends ModelProps> {
   [propsCreationDataSymbol]: ModelPropsToPropsCreationData<TProps>
   [instanceCreationDataSymbol]: ModelPropsToInstanceCreationData<TProps>
 
-  [composedInstanceCreationDataSymbol]: SuperModel extends BaseModel<any, infer ICD>
+  [composedInstanceCreationDataSymbol]: SuperModel extends BaseModel<any, any, any, infer ICD, any>
     ? this[typeof instanceCreationDataSymbol] & ICD
     : this[typeof instanceCreationDataSymbol]
 
-  new (
-    data: this[typeof composedInstanceCreationDataSymbol] & { [modelIdKey]?: string }
-  ): SuperModel &
+  new (data: this[typeof composedInstanceCreationDataSymbol]): SuperModel &
     BaseModel<
       this[typeof propsDataSymbol],
       this[typeof propsCreationDataSymbol],
       this[typeof instanceDataSymbol],
-      this[typeof instanceCreationDataSymbol]
+      this[typeof instanceCreationDataSymbol],
+      ExtractModelIdProp<TProps> & string
     > &
     Omit<this[typeof instanceDataSymbol], keyof AnyModel> &
     ModelPropsToSetterActions<TProps>
 }
+
+/**
+ * @ignore
+ * Ensures that a $modelId property is present if no idProp is provided.
+ */
+export type AddModelIdPropIfNeeded<
+  TProps extends ModelProps
+> = ExtractModelIdProp<TProps> extends never
+  ? TProps & { $modelId: typeof idProp } // we use the actual name here to avoid having to re-export the original
+  : TProps
+
+/**
+ * Extract the model id property from the model props.
+ */
+export type ExtractModelIdProp<TProps extends ModelProps> = {
+  [K in keyof TProps]: TProps[K]["$isId"] extends true ? K : never
+}[keyof TProps]
 
 /**
  * Base abstract class for models that extends another model.
@@ -78,11 +100,14 @@ export interface _Model<SuperModel, TProps extends ModelProps> {
 export function ExtendedModel<TProps extends ModelProps, TBaseModelClass>(
   baseModel: TBaseModelClass,
   modelProps: TProps
-): _Model<TBaseModelClass & Object extends ModelClass<infer M> ? M : never, TProps> {
+): _Model<
+  TBaseModelClass & Object extends ModelClass<infer M> ? M : never,
+  AddModelIdPropIfNeeded<TProps>
+> {
   assertIsModelClass(baseModel, "baseModel")
 
   // note that & Object is there to support abstract classes
-  return internalModel<TProps, any>(modelProps, baseModel as any)
+  return internalModel(modelProps, baseModel as any)
 }
 
 /**
@@ -93,14 +118,16 @@ export function ExtendedModel<TProps extends ModelProps, TBaseModelClass>(
  * @typeparam TProps Model properties type.
  * @param modelProps Model properties.
  */
-export function Model<TProps extends ModelProps>(modelProps: TProps): _Model<unknown, TProps> {
-  return internalModel(modelProps) as any
+export function Model<TProps extends ModelProps>(
+  modelProps: TProps
+): _Model<unknown, AddModelIdPropIfNeeded<TProps>> {
+  return internalModel(modelProps, undefined)
 }
 
 function internalModel<TProps extends ModelProps, TBaseModel extends AnyModel>(
   modelProps: TProps,
-  baseModel?: ModelClass<TBaseModel>
-): _Model<TBaseModel, TProps> {
+  baseModel: ModelClass<TBaseModel> | undefined
+): _Model<TBaseModel, AddModelIdPropIfNeeded<TProps>> {
   assertIsObject(modelProps, "modelProps")
   if (baseModel) {
     assertIsModelClass(baseModel, "baseModel")
@@ -126,14 +153,27 @@ function internalModel<TProps extends ModelProps, TBaseModel extends AnyModel>(
       }
       composedModelProps[oldModelPropKey] = oldModelProps[oldModelPropKey]
     }
-  } else {
-    // define $modelId on the base
-    extraDescriptors[modelIdKey] = createModelPropDescriptor(modelIdKey, undefined, true)
   }
+
+  // look for id keys
+  const idKeys = Object.keys(composedModelProps).filter((k) => composedModelProps[k] === idProp)
+  if (idKeys.length > 1) {
+    throw failure(`expected at most one idProp but got many: ${JSON.stringify(idKeys)}`)
+  }
+  if (idKeys.length <= 0) {
+    idKeys.push(modelIdKey)
+  }
+
+  const needsTypeChecker = Object.values(composedModelProps).some((mp) => !!mp.typeChecker)
+
+  // transform id keys (only one really)
+  const idKey = idKeys[0]
+  const idGenerator = () => getGlobalConfig().modelIdGenerator()
+  composedModelProps[idKey] = needsTypeChecker ? tProp(typesString, idGenerator) : prop(idGenerator)
 
   // create type checker if needed
   let dataTypeChecker: LateTypeChecker | undefined
-  if (Object.values(composedModelProps).some((mp) => !!mp.typeChecker)) {
+  if (needsTypeChecker) {
     const typeCheckerObj: {
       [k: string]: any
     } = {}
@@ -201,7 +241,11 @@ function internalModel<TProps extends ModelProps, TBaseModel extends AnyModel>(
   }
 
   setInternalModelClassPropsInfo(CustomBaseModel, composedModelProps)
-  CustomBaseModel[modelDataTypeCheckerSymbol] = dataTypeChecker
+
+  CustomBaseModel[modelMetadataSymbol] = {
+    dataType: dataTypeChecker,
+    modelIdProperty: idKey,
+  } as ModelMetadata
 
   Object.defineProperties(CustomBaseModel.prototype, extraDescriptors)