+++ packages/lib/src/model/modelDecorator.ts
@@ -68,7 +68,9 @@ const internalModel = (name: string) => <MC extends ModelClass<AnyModel>>(clazz:
         // was rejected on the mobx side
         if (
           err.message !==
-          "[MobX] No annotations were passed to makeObservable, but no decorator members have been found either"
+            "[MobX] No annotations were passed to makeObservable, but no decorator members have been found either" &&
+          err.message !==
+            "[MobX] No annotations were passed to makeObservable, but no decorated members have been found either"
         ) {
           throw err
         }