+++ packages/lib/src/model/newModel.ts
@@ -1,12 +1,12 @@
 import { action, set } from "mobx"
 import { O } from "ts-toolbelt"
-import { getGlobalConfig, isModelAutoTypeCheckingEnabled } from "../globalConfig/globalConfig"
+import { isModelAutoTypeCheckingEnabled } from "../globalConfig/globalConfig"
 import { tweakModel } from "../tweaker/tweakModel"
 import { tweakPlainObject } from "../tweaker/tweakPlainObject"
 import { failure, inDevMode, makePropReadonly } from "../utils"
 import { AnyModel, ModelPropsCreationData } from "./BaseModel"
-import { getModelDataType } from "./getModelDataType"
-import { modelIdKey, modelTypeKey } from "./metadata"
+import { getModelIdPropertyName, getModelMetadata } from "./getModelMetadata"
+import { modelTypeKey } from "./metadata"
 import { getModelClassInitializers } from "./modelClassInitializer"
 import { ModelConstructorOptions } from "./ModelConstructorOptions"
 import { modelInfoByClass } from "./modelInfo"
@@ -22,7 +22,7 @@ export const internalNewModel = action(
   "newModel",
   <M extends AnyModel>(
     origModelObj: M,
-    initialData: (ModelPropsCreationData<M> & { [modelIdKey]?: string }) | undefined,
+    initialData: ModelPropsCreationData<M> | undefined,
     options: Pick<ModelConstructorOptions, "modelClass" | "snapshotInitialData" | "generateNewIds">
   ): M => {
     const { modelClass: _modelClass, snapshotInitialData, generateNewIds } = options
@@ -41,14 +41,18 @@ export const internalNewModel = action(
       )
     }
 
+    const modelIdPropertyName = getModelIdPropertyName(modelClass)
+    const modelProps = getInternalModelClassPropsInfo(modelClass)
+    const modelIdPropData = modelProps[modelIdPropertyName]!
+
     let id
     if (snapshotInitialData) {
       let sn = snapshotInitialData.unprocessedSnapshot
 
       if (generateNewIds) {
-        id = getGlobalConfig().modelIdGenerator()
+        id = (modelIdPropData.defaultFn as () => string)()
       } else {
-        id = sn[modelIdKey]
+        id = sn[modelIdPropertyName]
       }
 
       if (modelObj.fromSnapshot) {
@@ -58,34 +62,36 @@ export const internalNewModel = action(
       initialData = snapshotInitialData.snapshotToInitialData(sn)
     } else {
       // use symbol if provided
-      if (initialData![modelIdKey]) {
-        id = initialData![modelIdKey]
+      if (initialData![modelIdPropertyName]) {
+        id = initialData![modelIdPropertyName]
       } else {
-        id = getGlobalConfig().modelIdGenerator()
+        id = (modelIdPropData.defaultFn as () => string)()
       }
     }
 
     modelObj[modelTypeKey] = modelInfo.name
 
     // fill in defaults in initial data
-    const modelProps = getInternalModelClassPropsInfo(modelClass)
     const modelPropsKeys = Object.keys(modelProps)
     for (let i = 0; i < modelPropsKeys.length; i++) {
       const k = modelPropsKeys[i]
-      const v = (initialData as any)[k]
-      if (v === undefined || v === null) {
-        let newValue: any = v
-        const propData = modelProps[k]
-        if (propData.defaultFn !== noDefaultValue) {
-          newValue = propData.defaultFn()
-        } else if (propData.defaultValue !== noDefaultValue) {
-          newValue = propData.defaultValue
+      // id is already initialized above
+      if (k !== modelIdPropertyName) {
+        const v = (initialData as any)[k]
+        if (v === undefined || v === null) {
+          let newValue: any = v
+          const propData = modelProps[k]
+          if (propData.defaultFn !== noDefaultValue) {
+            newValue = propData.defaultFn()
+          } else if (propData.defaultValue !== noDefaultValue) {
+            newValue = propData.defaultValue
+          }
+          set(initialData as any, k, newValue)
         }
-        set(initialData as any, k, newValue)
       }
     }
 
-    set(initialData as any, modelIdKey, id)
+    set(initialData as any, modelIdPropertyName, id)
 
     tweakModel(modelObj, undefined)
 
@@ -98,12 +104,6 @@ export const internalNewModel = action(
       true
     )
 
-    // hide $.$modelId
-    Object.defineProperty(obsData, modelIdKey, {
-      ...Object.getOwnPropertyDescriptor(obsData, modelIdKey),
-      enumerable: false,
-    })
-
     // link it, and make it readonly
     modelObj.$ = obsData
     if (inDevMode()) {
@@ -111,7 +111,7 @@ export const internalNewModel = action(
     }
 
     // type check it if needed
-    if (isModelAutoTypeCheckingEnabled() && getModelDataType(modelClass)) {
+    if (isModelAutoTypeCheckingEnabled() && getModelMetadata(modelClass).dataType) {
       const err = modelObj.typeCheck()
       if (err) {
         err.throw(modelObj)