+++ packages/lib/src/typeChecking/typeCheck.ts
@@ -5,7 +5,7 @@ import { TypeCheckError } from "./TypeCheckError"
 /**
  * Checks if a value conforms to a given type.
  *
- * @typename S Type.
+ * @typename T Type.
  * @param type Type to check for.
  * @param value Value to check.
  * @returns A TypeError if the check fails or null if no error.