+++ packages/lib/test/actionMiddlewares/undoMiddleware.test.ts
@@ -13,6 +13,7 @@ import {
   UndoManager,
   undoMiddleware,
   UndoStore,
+  withoutUndo,
   _async,
   _await,
 } from "../../src"
@@ -60,6 +61,23 @@ class R extends Model({
   p: prop(() => new P({})),
 }) {}
 
+function expectUndoManagerRedoToBe(manager: UndoManager, undoLevels: number, redoLevels: number) {
+  expect(manager.canUndo).toBe(undoLevels > 0)
+  expect(manager.undoLevels).toBe(undoLevels)
+  expect(manager.undoQueue.length).toBe(undoLevels)
+
+  expect(manager.canRedo).toBe(redoLevels > 0)
+  expect(manager.redoLevels).toBe(redoLevels)
+  expect(manager.redoQueue.length).toBe(redoLevels)
+
+  if (undoLevels <= 0) {
+    expect(() => manager.undo()).toThrow("nothing to undo")
+  }
+  if (redoLevels <= 0) {
+    expect(() => manager.redo()).toThrow("nothing to redo")
+  }
+}
+
 test("undoMiddleware - sync", () => {
   const r = new R({})
   const p = r.p
@@ -76,20 +94,7 @@ test("undoMiddleware - sync", () => {
   }
 
   function expectUndoRedoToBe(undoLevels: number, redoLevels: number) {
-    expect(manager.canUndo).toBe(undoLevels > 0)
-    expect(manager.undoLevels).toBe(undoLevels)
-    expect(manager.undoQueue.length).toBe(undoLevels)
-
-    expect(manager.canRedo).toBe(redoLevels > 0)
-    expect(manager.redoLevels).toBe(redoLevels)
-    expect(manager.redoQueue.length).toBe(redoLevels)
-
-    if (undoLevels <= 0) {
-      expect(() => manager.undo()).toThrow("nothing to undo")
-    }
-    if (redoLevels <= 0) {
-      expect(() => manager.redo()).toThrow("nothing to redo")
-    }
+    expectUndoManagerRedoToBe(manager, undoLevels, redoLevels)
   }
 
   expectUndoRedoToBe(0, 0)
@@ -143,6 +148,7 @@ test("undoMiddleware - sync", () => {
           "targetPath": Array [
             "p",
           ],
+          "type": "single",
         },
         Object {
           "actionName": "incX",
@@ -169,6 +175,7 @@ test("undoMiddleware - sync", () => {
           "targetPath": Array [
             "p",
           ],
+          "type": "single",
         },
         Object {
           "actionName": "incY",
@@ -198,6 +205,7 @@ test("undoMiddleware - sync", () => {
             "p",
             "p2",
           ],
+          "type": "single",
         },
         Object {
           "actionName": "incXY",
@@ -242,6 +250,7 @@ test("undoMiddleware - sync", () => {
           "targetPath": Array [
             "p",
           ],
+          "type": "single",
         },
       ],
     }
@@ -329,6 +338,7 @@ test("undoMiddleware - sync", () => {
         "targetPath": Array [
           "p",
         ],
+        "type": "single",
       },
       Object {
         "actionName": "pushArr",
@@ -375,6 +385,7 @@ test("undoMiddleware - sync", () => {
         "targetPath": Array [
           "p",
         ],
+        "type": "single",
       },
     ]
   `)
@@ -465,20 +476,7 @@ test("undoMiddleware - async", async () => {
   }
 
   function expectUndoRedoToBe(undoLevels: number, redoLevels: number) {
-    expect(manager.canUndo).toBe(undoLevels > 0)
-    expect(manager.undoLevels).toBe(undoLevels)
-    expect(manager.undoQueue.length).toBe(undoLevels)
-
-    expect(manager.canRedo).toBe(redoLevels > 0)
-    expect(manager.redoLevels).toBe(redoLevels)
-    expect(manager.redoQueue.length).toBe(redoLevels)
-
-    if (undoLevels <= 0) {
-      expect(() => manager.undo()).toThrow("nothing to undo")
-    }
-    if (redoLevels <= 0) {
-      expect(() => manager.redo()).toThrow("nothing to redo")
-    }
+    expectUndoManagerRedoToBe(manager, undoLevels, redoLevels)
   }
 
   expectUndoRedoToBe(0, 0)
@@ -537,6 +535,7 @@ test("undoMiddleware - async", async () => {
           "targetPath": Array [
             "p",
           ],
+          "type": "single",
         },
         Object {
           "actionName": "incX",
@@ -563,6 +562,7 @@ test("undoMiddleware - async", async () => {
           "targetPath": Array [
             "p",
           ],
+          "type": "single",
         },
         Object {
           "actionName": "incY",
@@ -592,6 +592,7 @@ test("undoMiddleware - async", async () => {
             "p",
             "p2",
           ],
+          "type": "single",
         },
         Object {
           "actionName": "incXY",
@@ -636,6 +637,7 @@ test("undoMiddleware - async", async () => {
           "targetPath": Array [
             "p",
           ],
+          "type": "single",
         },
       ],
     }
@@ -830,6 +832,7 @@ test("undo-aware substore called from non undo-aware root store", () => {
         "targetPath": Array [
           "substore",
         ],
+        "type": "single",
       },
     ]
   `)
@@ -871,8 +874,443 @@ test("undo-aware substore called from non undo-aware root store", () => {
         "targetPath": Array [
           "substore",
         ],
+        "type": "single",
       },
     ]
   `)
   manager.clearUndo()
+})
+
+test("does not generate steps if using withoutUndo", () => {
+  const r = new R({})
+  const p = r.p
+
+  const manager = undoMiddleware(r, r.undoData)
+  autoDispose(() => manager.dispose())
+
+  function expectUndoRedoToBe(undoLevels: number, redoLevels: number) {
+    expectUndoManagerRedoToBe(manager, undoLevels, redoLevels)
+  }
+
+  expectUndoRedoToBe(0, 0)
+
+  manager.withoutUndo(() => {
+    p.incX(1)
+    p.incX(1)
+  })
+
+  withoutUndo(() => {
+    p.incX(1)
+    p.incX(1)
+  })
+
+  expect(p.x).toBe(4)
+  expectUndoRedoToBe(0, 0)
+})
+
+test("withGroup", () => {
+  const r = new R({})
+  const p = r.p
+
+  const manager = undoMiddleware(r, r.undoData)
+  autoDispose(() => manager.dispose())
+
+  function expectUndoRedoToBe(undoLevels: number, redoLevels: number) {
+    expectUndoManagerRedoToBe(manager, undoLevels, redoLevels)
+  }
+
+  expectUndoRedoToBe(0, 0)
+
+  manager.withGroup("group1", () => {
+    p.incX(1)
+    p.incX(2)
+
+    manager.withGroup(() => {
+      p.incX(3)
+      p.incX(4)
+    })
+  })
+
+  expect(p.x).toBe(10)
+  expectUndoRedoToBe(1, 0)
+  expect(manager.undoQueue).toMatchInlineSnapshot(`
+    Array [
+      Object {
+        "events": Array [
+          Object {
+            "actionName": "incX",
+            "inversePatches": Array [
+              Object {
+                "op": "replace",
+                "path": Array [
+                  "p",
+                  "x",
+                ],
+                "value": 0,
+              },
+            ],
+            "patches": Array [
+              Object {
+                "op": "replace",
+                "path": Array [
+                  "p",
+                  "x",
+                ],
+                "value": 1,
+              },
+            ],
+            "targetPath": Array [
+              "p",
+            ],
+            "type": "single",
+          },
+          Object {
+            "actionName": "incX",
+            "inversePatches": Array [
+              Object {
+                "op": "replace",
+                "path": Array [
+                  "p",
+                  "x",
+                ],
+                "value": 1,
+              },
+            ],
+            "patches": Array [
+              Object {
+                "op": "replace",
+                "path": Array [
+                  "p",
+                  "x",
+                ],
+                "value": 3,
+              },
+            ],
+            "targetPath": Array [
+              "p",
+            ],
+            "type": "single",
+          },
+          Object {
+            "events": Array [
+              Object {
+                "actionName": "incX",
+                "inversePatches": Array [
+                  Object {
+                    "op": "replace",
+                    "path": Array [
+                      "p",
+                      "x",
+                    ],
+                    "value": 3,
+                  },
+                ],
+                "patches": Array [
+                  Object {
+                    "op": "replace",
+                    "path": Array [
+                      "p",
+                      "x",
+                    ],
+                    "value": 6,
+                  },
+                ],
+                "targetPath": Array [
+                  "p",
+                ],
+                "type": "single",
+              },
+              Object {
+                "actionName": "incX",
+                "inversePatches": Array [
+                  Object {
+                    "op": "replace",
+                    "path": Array [
+                      "p",
+                      "x",
+                    ],
+                    "value": 6,
+                  },
+                ],
+                "patches": Array [
+                  Object {
+                    "op": "replace",
+                    "path": Array [
+                      "p",
+                      "x",
+                    ],
+                    "value": 10,
+                  },
+                ],
+                "targetPath": Array [
+                  "p",
+                ],
+                "type": "single",
+              },
+            ],
+            "groupName": undefined,
+            "type": "group",
+          },
+        ],
+        "groupName": "group1",
+        "type": "group",
+      },
+    ]
+  `)
+
+  manager.undo()
+  expectUndoRedoToBe(0, 1)
+  expect(p.x).toBe(0)
+  expect(manager.redoQueue).toMatchInlineSnapshot(`
+    Array [
+      Object {
+        "events": Array [
+          Object {
+            "actionName": "incX",
+            "inversePatches": Array [
+              Object {
+                "op": "replace",
+                "path": Array [
+                  "p",
+                  "x",
+                ],
+                "value": 0,
+              },
+            ],
+            "patches": Array [
+              Object {
+                "op": "replace",
+                "path": Array [
+                  "p",
+                  "x",
+                ],
+                "value": 1,
+              },
+            ],
+            "targetPath": Array [
+              "p",
+            ],
+            "type": "single",
+          },
+          Object {
+            "actionName": "incX",
+            "inversePatches": Array [
+              Object {
+                "op": "replace",
+                "path": Array [
+                  "p",
+                  "x",
+                ],
+                "value": 1,
+              },
+            ],
+            "patches": Array [
+              Object {
+                "op": "replace",
+                "path": Array [
+                  "p",
+                  "x",
+                ],
+                "value": 3,
+              },
+            ],
+            "targetPath": Array [
+              "p",
+            ],
+            "type": "single",
+          },
+          Object {
+            "events": Array [
+              Object {
+                "actionName": "incX",
+                "inversePatches": Array [
+                  Object {
+                    "op": "replace",
+                    "path": Array [
+                      "p",
+                      "x",
+                    ],
+                    "value": 3,
+                  },
+                ],
+                "patches": Array [
+                  Object {
+                    "op": "replace",
+                    "path": Array [
+                      "p",
+                      "x",
+                    ],
+                    "value": 6,
+                  },
+                ],
+                "targetPath": Array [
+                  "p",
+                ],
+                "type": "single",
+              },
+              Object {
+                "actionName": "incX",
+                "inversePatches": Array [
+                  Object {
+                    "op": "replace",
+                    "path": Array [
+                      "p",
+                      "x",
+                    ],
+                    "value": 6,
+                  },
+                ],
+                "patches": Array [
+                  Object {
+                    "op": "replace",
+                    "path": Array [
+                      "p",
+                      "x",
+                    ],
+                    "value": 10,
+                  },
+                ],
+                "targetPath": Array [
+                  "p",
+                ],
+                "type": "single",
+              },
+            ],
+            "groupName": undefined,
+            "type": "group",
+          },
+        ],
+        "groupName": "group1",
+        "type": "group",
+      },
+    ]
+  `)
+
+  manager.redo()
+  expectUndoRedoToBe(1, 0)
+  expect(p.x).toBe(10)
+  expect(manager.undoQueue).toMatchInlineSnapshot(`
+    Array [
+      Object {
+        "events": Array [
+          Object {
+            "actionName": "incX",
+            "inversePatches": Array [
+              Object {
+                "op": "replace",
+                "path": Array [
+                  "p",
+                  "x",
+                ],
+                "value": 0,
+              },
+            ],
+            "patches": Array [
+              Object {
+                "op": "replace",
+                "path": Array [
+                  "p",
+                  "x",
+                ],
+                "value": 1,
+              },
+            ],
+            "targetPath": Array [
+              "p",
+            ],
+            "type": "single",
+          },
+          Object {
+            "actionName": "incX",
+            "inversePatches": Array [
+              Object {
+                "op": "replace",
+                "path": Array [
+                  "p",
+                  "x",
+                ],
+                "value": 1,
+              },
+            ],
+            "patches": Array [
+              Object {
+                "op": "replace",
+                "path": Array [
+                  "p",
+                  "x",
+                ],
+                "value": 3,
+              },
+            ],
+            "targetPath": Array [
+              "p",
+            ],
+            "type": "single",
+          },
+          Object {
+            "events": Array [
+              Object {
+                "actionName": "incX",
+                "inversePatches": Array [
+                  Object {
+                    "op": "replace",
+                    "path": Array [
+                      "p",
+                      "x",
+                    ],
+                    "value": 3,
+                  },
+                ],
+                "patches": Array [
+                  Object {
+                    "op": "replace",
+                    "path": Array [
+                      "p",
+                      "x",
+                    ],
+                    "value": 6,
+                  },
+                ],
+                "targetPath": Array [
+                  "p",
+                ],
+                "type": "single",
+              },
+              Object {
+                "actionName": "incX",
+                "inversePatches": Array [
+                  Object {
+                    "op": "replace",
+                    "path": Array [
+                      "p",
+                      "x",
+                    ],
+                    "value": 6,
+                  },
+                ],
+                "patches": Array [
+                  Object {
+                    "op": "replace",
+                    "path": Array [
+                      "p",
+                      "x",
+                    ],
+                    "value": 10,
+                  },
+                ],
+                "targetPath": Array [
+                  "p",
+                ],
+                "type": "single",
+              },
+            ],
+            "groupName": undefined,
+            "type": "group",
+          },
+        ],
+        "groupName": "group1",
+        "type": "group",
+      },
+    ]
+  `)
 })